chore: Update Dockerfile and docker-compose files
This commit updates the Dockerfile and docker-compose files. The Dockerfile is modified to improve the build process by copying the entire project into the container and setting the working directory. The docker-compose files are removed as they are no longer needed. These changes optimize the deployment and containerization of the ASP.NET Core service.